feat: add tell/tellI
Add the following functions:

- TaskWriterEither
  - tellI
- ReaderTaskWriterEither
  - tell
  - tellI

We take this opportunity to add a new suffix convention, like `W` for
`widen` or `K` for `kleisli`: `I`, for `identity`.

Instead of the following pattern:

```typescript
pipe(
  TWE.tryCatch('launching missiles')(launchMissiles, (error) => error),
  TWE.chain((result) => TWE.of(`result: ${result`)(result)),
)
```

(in other words, a `tell` where we want to use the value of `a` in the
log), we can indicate that `chain` returns the given value (aka the
`identity` function) with `tellI`:

```typescript
pipe(
  TWE.tryCatch('launching missiles')(launchMissiles, (error) => error),
  TWE.tellI((result) => `result: ${result`),
)
```

I'm not sure if there's a pattern for this in more functional languages,
or if I'm using `tell` incorrectly, but this pattern will be useful for
explicit logs so we will give it a shot. Please drop me a line if this
is an identifiable anti-pattern.