perf: mutate log when adding a single entry
for better performance. This is most certainly a hack but allows the log
to grow in amortized-doubling O(n) time rather than O(n) every time.
This should make the `Writer` monad usable in more situations.

The risk of mutating values is kept to a minimum by not exposing that
this is in fact a mutable log under-the-hood.